Use both the subdomains and subdomains-aggregate endpoints
The findsubdomains.com API has two endpoints related subdomains, that each give a subset of the total data.

Call both of them to get as much data as possible.

The subdomains endpoint requires credit to query more than the first page, so it is only guaranteed to give you 30 domains, so as a fallback we can query the subdomains-aggregate endpoint too.